# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gig Trail Tour Calculator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles:
    get:
      operationId: getVehicles
      tags: [vehicles]
      summary: List all vehicles
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Create a vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleBody"
      responses:
        "201":
          description: Created vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      summary: Get a vehicle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Update a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleBody"
      responses:
        "200":
          description: Updated vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    delete:
      operationId: deleteVehicle
      tags: [vehicles]
      summary: Delete a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /tours:
    get:
      operationId: getTours
      tags: [tours]
      summary: List all tours
      responses:
        "200":
          description: List of tours
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tour"
    post:
      operationId: createTour
      tags: [tours]
      summary: Create a tour
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTourBody"
      responses:
        "201":
          description: Created tour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
  /tours/{id}:
    get:
      operationId: getTour
      tags: [tours]
      summary: Get a tour with all its stops
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tour with stops
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourWithStops"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTour
      tags: [tours]
      summary: Update a tour
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTourBody"
      responses:
        "200":
          description: Updated tour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
    delete:
      operationId: deleteTour
      tags: [tours]
      summary: Delete a tour
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tours/{tourId}/stops:
    get:
      operationId: getTourStops
      tags: [tours]
      summary: List stops for a tour
      parameters:
        - name: tourId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of stops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TourStop"
    post:
      operationId: createTourStop
      tags: [tours]
      summary: Add a stop to a tour
      parameters:
        - name: tourId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTourStopBody"
      responses:
        "201":
          description: Created stop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourStop"
  /tours/{tourId}/stops/{stopId}:
    patch:
      operationId: updateTourStop
      tags: [tours]
      summary: Update a tour stop
      parameters:
        - name: tourId
          in: path
          required: true
          schema:
            type: integer
        - name: stopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTourStopBody"
      responses:
        "200":
          description: Updated stop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourStop"
    delete:
      operationId: deleteTourStop
      tags: [tours]
      summary: Delete a tour stop
      parameters:
        - name: tourId
          in: path
          required: true
          schema:
            type: integer
        - name: stopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        fuelType:
          type: string
        avgConsumption:
          type: number
        maxPassengers:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - fuelType
        - avgConsumption
        - maxPassengers
        - notes
        - createdAt
    CreateVehicleBody:
      type: object
      properties:
        name:
          type: string
        fuelType:
          type: string
        avgConsumption:
          type: number
        maxPassengers:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
        - fuelType
        - avgConsumption
...
    Tour:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        profileId:
          type: ["integer", "null"]
        vehicleId:
          type: ["integer", "null"]
        startLocation:
          type: ["string", "null"]
        endLocation:
          type: ["string", "null"]
        returnHome:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        defaultFoodCost:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        totalDistance:
          type: ["number", "null"]
        totalCost:
          type: ["number", "null"]
        totalIncome:
          type: ["number", "null"]
        totalProfit:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - profileId
        - vehicleId
        - startLocation
        - endLocation
        - returnHome
        - startDate
        - endDate
        - defaultFoodCost
        - notes
        - totalDistance
        - totalCost
        - totalIncome
        - totalProfit
        - createdAt
    CreateTourBody:
      type: object
      properties:
        name:
          type: string
        profileId:
          type: ["integer", "null"]
        vehicleId:
          type: ["integer", "null"]
        startLocation:
          type: ["string", "null"]
        endLocation:
          type: ["string", "null"]
        returnHome:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        defaultFoodCost:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
    TourStop:
      type: object
      properties:
        id:
          type: integer
        tourId:
          type: integer
        stopOrder:
          type: integer
        date:
          type: ["string", "null"]
        city:
          type: string
        venueName:
          type: ["string", "null"]
        showType:
          type: string
        fee:
          type: ["number", "null"]
        capacity:
          type: ["integer", "null"]
        ticketPrice:
          type: ["number", "null"]
        expectedAttendancePct:
          type: ["number", "null"]
        dealType:
          type: ["string", "null"]
        splitPct:
          type: ["number", "null"]
        guarantee:
          type: ["number", "null"]
        merchEstimate:
          type: ["number", "null"]
        marketingCost:
          type: ["number", "null"]
        accommodationCost:
          type: ["number", "null"]
        extraCosts:
          type: ["number", "null"]
        distanceOverride:
          type: ["number", "null"]
        fuelPriceOverride:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - id
        - tourId
        - stopOrder
        - date
        - city
        - venueName
        - showType
        - fee
        - capacity
        - ticketPrice
        - expectedAttendancePct
        - dealType
        - splitPct
        - guarantee
        - merchEstimate
        - marketingCost
        - accommodationCost
        - extraCosts
        - distanceOverride
        - fuelPriceOverride
        - notes
    CreateTourStopBody:
      type: object
      properties:
        stopOrder:
          type: integer
        date:
          type: ["string", "null"]
        city:
          type: string
        venueName:
          type: ["string", "null"]
        showType:
          type: string
        fee:
          type: ["number", "null"]
        capacity:
          type: ["integer", "null"]
        ticketPrice:
          type: ["number", "null"]
        expectedAttendancePct:
          type: ["number", "null"]
        dealType:
          type: ["string", "null"]
        splitPct:
          type: ["number", "null"]
        guarantee:
          type: ["number", "null"]
        merchEstimate:
          type: ["number", "null"]
        marketingCost:
          type: ["number", "null"]
        accommodationCost:
          type: ["number", "null"]
        extraCosts:
          type: ["number", "null"]
        distanceOverride:
          type: ["number", "null"]
        fuelPriceOverride:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - city
        - showType
    TourWithStops:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        profileId:
          type: ["integer", "null"]
        vehicleId:
          type: ["integer", "null"]
        startLocation:
          type: ["string", "null"]
        endLocation:
          type: ["string", "null"]
        returnHome:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        defaultFoodCost:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        totalDistance:
          type: ["number", "null"]
        totalCost:
          type: ["number", "null"]
        totalIncome:
          type: ["number", "null"]
        totalProfit:
          type: ["number", "null"]
        createdAt:
          type: string
        stops:
          type: array
          items:
            $ref: "#/components/schemas/TourStop"
      required:
        - id
        - name
        - profileId
        - vehicleId
        - startLocation
        - endLocation
        - returnHome
        - startDate
        - endDate
        - defaultFoodCost
        - notes
        - totalDistance
        - totalCost
        - totalIncome
        - totalProfit
        - createdAt
        - stops
...