# lib/api-zod/src/generated/types/createVehicleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gig Trail Tour Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateVehicleBody {
  name: string;
  fuelType: string;
  avgConsumption: number;
  /** @nullable */
  maxPassengers?: number | null;
  /** @nullable */
  notes?: string | null;
}