# lib/api-zod/src/generated/types/tourStop.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gig Trail Tour Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface TourStop {
  id: number;
  tourId: number;
  stopOrder: number;
  /** @nullable */
  date: string | null;
  city: string;
  /** @nullable */
  venueName: string | null;
  showType: string;
  /** @nullable */
  fee: number | null;
  /** @nullable */
  capacity: number | null;
  /** @nullable */
  ticketPrice: number | null;
  /** @nullable */
  expectedAttendancePct: number | null;
  /** @nullable */
  dealType: string | null;
  /** @nullable */
  splitPct: number | null;
  /** @nullable */
  guarantee: number | null;
  /** @nullable */
  merchEstimate: number | null;
  /** @nullable */
  marketingCost: number | null;
  /** @nullable */
  accommodationCost: number | null;
  /** @nullable */
  extraCosts: number | null;
  /** @nullable */
  distanceOverride: number | null;
  /** @nullable */
  fuelPriceOverride: number | null;
  /** @nullable */
  notes: string | null;
}