# lib/api-zod/src/generated/types/tourWithStops.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gig Trail Tour Calculator API
 * OpenAPI spec version: 0.1.0
 */
import type { TourStop } from "./tourStop";

export interface TourWithStops {
  id: number;
  name: string;
  /** @nullable */
  profileId: number | null;
  /** @nullable */
  vehicleId: number | null;
  /** @nullable */
  startLocation: string | null;
  /** @nullable */
  endLocation: string | null;
  returnHome: boolean;
  /** @nullable */
  startDate: string | null;
  /** @nullable */
  endDate: string | null;
  /** @nullable */
  defaultFoodCost: number | null;
  /** @nullable */
  notes: string | null;
  /** @nullable */
  totalDistance: number | null;
  /** @nullable */
  totalCost: number | null;
  /** @nullable */
  totalIncome: number | null;
  /** @nullable */
  totalProfit: number | null;
  createdAt: string;
  stops: TourStop[];
}