# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gig Trail Tour Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface Vehicle {
  id: number;
  name: string;
  fuelType: string;
  avgConsumption: number;
  /** @nullable */
  maxPassengers: number | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}